requirements.txt
asgiref==3.11.1
boto3==1.42.52
botocore==1.42.52
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.2
django-allauth==65.14.3
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
jmespath==1.1.0
mercadopago==2.3.0
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.10
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
stripe==14.3.0
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
